Update list examples & nightwatch tests.
