+++ packages/terra-list/tests/nightwatch/list/ListTests.jsx
@@ -7,7 +7,6 @@ const ListTests = () => (
   <div>
     <ul>
       <li><Link to="/tests/list-tests/default">Default List</Link></li>
-      <li><Link to="/tests/list-tests/items">Items List</Link></li>
       <li><Link to="/tests/list-tests/items-divided">Items Divided List</Link></li>
     </ul>
   </div>
